requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.5
astunparse>=1.6.3
pytest>=7.4.3
docker>=6.1.3